[FRCV-16] Improvements - Part 1 (#21)
* [FRCV-17] Organize the types into a one main folder (#16)
Organized all type files under the main types folder.

* [FRCV-19] Improve user feedback for waiting ai response (#17)
Sending feedback to user if timeout 1.5s and 5s

* [FRCV-18] Improve adding types where is still as any (#18)
Added more types where it was needed

* [FRCV-20] Add docstrings (#19)
Added the docstring to the project

* [FRCV-21] Create a README description (#20)
Created the readme for the project

* [FRCV-16] Fixing the readme and a bug
I noticed a bug when reviewing the readme, so it's also contain a small fix.