[v1.0.0] Release - Developing the first version (#22)
* [FRCV-1] Cluster Implementation (#5)

* [FRCV-4] Install cluster dependencies (#1)
Installing the basic for implementing a cluster

* [FRCV-5] Cluster Service (#2)
Created the service to manage the CLUSTER and its CPUs/Thread under.

* [FRCV-6] Install the Slack dependency (#3)
Installed the Bolt framework to handle the slack apps

* [FRCV-8] SlackApp CPU/Thread (#4)
Creating the slack-app core and the slack-main thread.

* [FRCV-3] Slack App (#8)

* [FRCV-7] SlackApp Service (#6)
Created the base for SlackApp instance to manage slack features

* [FRCV-9] SlackApp listeners (#7)
Created the methods to implement the listeners for message and action

* [FRCV-2] OpenAI (#12)

* [FRCV-10] Create the main Thread for OpenAI (#9)
Created the main thread and core to be used by the open ai integration

* [FRCV-11] AI Service (#10)
Created the main AI instance to handle integration with OpenAI API

* [FRCV-12] EventEndpoints Callback (#11)
Created the callback feature for the Event endpoints, so we can provide responses through the instances.

* Update package.json

* [FRCV-13] GPT Replying Slack (#15)

* [FRCV-15] AI Endpoints (#13)
Created the communication between the slack app and the ai assistant

* [FRCV-14] Error Builder (#14)
Created the ErrorModel to build errors and stored as global to be used across the app.

* [FRCV-16] Improvements - Part 1 (#21)

* [FRCV-17] Organize the types into a one main folder (#16)
Organized all type files under the main types folder.

* [FRCV-19] Improve user feedback for waiting ai response (#17)
Sending feedback to user if timeout 1.5s and 5s

* [FRCV-18] Improve adding types where is still as any (#18)
Added more types where it was needed

* [FRCV-20] Add docstrings (#19)
Added the docstring to the project

* [FRCV-21] Create a README description (#20)
Created the readme for the project

* [FRCV-16] Fixing the readme and a bug
I noticed a bug when reviewing the readme, so it's also contain a small fix.