[FRCV-146] Edit and Delete Opportunity
## [FRCV-146] Description
Introduces PATCH and DELETE endpoints for Opportunity, allowing admins and masters to update or delete opportunities. Adds Opportunity model methods for updating and deleting, including optional deletion of related CV and Company. Also fixes a typo in the create route (jobUrl to jobURL) and updates type definitions for search parameters.